restructure HTML and redo Footer
