+++ app/components/icons/RaiderIO.tsx
@@ -1,0 +1,21 @@
+import clsx from "clsx";
+import { type SVGProps } from "react";
+
+export default function RaiderIO({
+  className,
+  ...props
+}: SVGProps<SVGSVGElement>): JSX.Element {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      aria-hidden="true"
+      viewBox="0 0 300 300"
+      className={clsx("h-6 w-6", className)}
+      fill="#fff"
+      {...props}
+    >
+      <path d="M294.91 117.4a8.75 8.75 0 0 0 2.22-7.3c-.21-.8-.4-1.6-.63-2.4a8.93 8.93 0 0 0-5.68-5.31l-27.93-8.3a13.45 13.45 0 0 1-9.38-16l6.66-27.94a8.9 8.9 0 0 0-1-6.71A9.36 9.36 0 0 0 249 39.19l-28.41 6.55a13.73 13.73 0 0 1-16.24-9.22l-8.44-27.47a8.89 8.89 0 0 0-5.27-5.52c-.91-.25-1.82-.46-2.74-.69A9 9 0 0 0 180.65 5l-21.27 19.67a13.89 13.89 0 0 1-18.76 0L119.35 5a9 9 0 0 0-7.25-2.19c-.91.23-1.83.44-2.74.69a8.89 8.89 0 0 0-5.27 5.52l-8.44 27.5a13.73 13.73 0 0 1-16.24 9.22L51 39.19a9.36 9.36 0 0 0-10.13 4.28 8.9 8.9 0 0 0-1 6.71l6.66 27.94a13.45 13.45 0 0 1-9.38 16l-27.93 8.3a8.89 8.89 0 0 0-5.68 5.32c-.23.79-.42 1.59-.63 2.39a8.72 8.72 0 0 0 2.22 7.3l20 20.92a13.31 13.31 0 0 1 0 18.44l-20 20.93a8.72 8.72 0 0 0-2.22 7.4c.19.74.37 1.47.57 2.2a9 9 0 0 0 5.74 5.4L37.11 201a13.44 13.44 0 0 1 9.38 16l-6.66 27.95a8.88 8.88 0 0 0 1.45 7.26 9.54 9.54 0 0 0 2.3 2.27 9.23 9.23 0 0 0 7.42 1.41l47-5a4.32 4.32 0 0 0 3.92-4.67c-1.57-16.95-8.93-80.33-11.7-104A17.51 17.51 0 0 1 96 127.13l46-41.23a12 12 0 0 1 16 0l46 41.23a17.53 17.53 0 0 1 5.78 15.08c-2.74 23.64-10.06 87.05-11.62 104a4.32 4.32 0 0 0 3.92 4.67l47 5a9.42 9.42 0 0 0 9.71-3.72 8.84 8.84 0 0 0 1.46-7.26l-6.74-27.9a13.46 13.46 0 0 1 9.38-16l27.93-8.3a8.9 8.9 0 0 0 5.73-5.4c.21-.73.39-1.47.58-2.2a8.76 8.76 0 0 0-2.22-7.4l-20-20.93a13.32 13.32 0 0 1 0-18.44Z" />
+      <path d="M154.55 124.05a6.09 6.09 0 0 0-9.1 0l-15.77 17.84a17.93 17.93 0 0 0-4.49 13l8 133.36a9.57 9.57 0 0 0 9.23 8.82q3.75.2 7.54.19t7.59-.19a9.56 9.56 0 0 0 9.23-8.83l8-133.34a17.93 17.93 0 0 0-4.5-13Z" />
+    </svg>
+  );
+}
