+++ package.json
@@ -25,6 +25,7 @@
     "@vercel/analytics": "0.1.11",
     "@vercel/remix-entry-server": "0.1.1",
     "cheerio": "1.0.0-rc.10",
+    "clsx": "1.2.1",
     "highcharts": "10.3.3",
     "highcharts-react-official": "3.2.0",
     "react": "18.2.0",
